Add two view to users edit and permission, add image input to registration form, create edit form for user

+++ routes/web.php
@@ -31,32 +31,27 @@ Route::get('/dashboard', function () {
 
 
 Route::middleware(['auth', 'role:admin'])->name('admin.')->prefix('admin')->group(function () {
-  
+
     Route::get('/', [IndexController::class, 'index'])->name('index');
 
 
     Route::resource('/roles', RoleController::class);
     Route::post('/roles/{role}/permissions', [RoleController::class, 'givePermission'])->name('roles.permissions');
     Route::delete('/roles/{role}/permissions/{permission}', [RoleController::class, 'revokePermission'])->name('roles.permissions.revoke');
-   
-   
+
+
     Route::resource('/permissions', PermissionController::class);
     Route::post('/permissions/{permission}/roles', [PermissionController::class, 'assignRole'])->name('permissions.roles');
     Route::delete('/permissions/{permission}/roles/{role}', [PermissionController::class, 'removeRole'])->name('permissions.roles.remove');
-   
-   
-    // Route::get('/users', [UserController::class, 'index'])->name('users.index');
-    // Route::post('/users/create', [UserController::class, 'create'])->name('users.create');
-   
     Route::resource('/users',UserController::class);
-
     Route::get('/users/{user}', [UserController::class, 'show'])->name('users.show');
-    Route::delete('/users/{user}', [UserController::class, 'destroy'])->name('users.destroy');
-   
+    Route::get('/users/{user}/permission', [UserController::class, 'permission'])->name('users.permission');
+    // Route::get('/users/{user}/edit', [UserController::class, 'edit'])->name('users.edit');
+    // Route::put('/users/{user}/edit/update', [UserController::class, 'update'])->name('users.update');
     Route::post('/users/{user}/roles', [UserController::class, 'assignRole'])->name('users.roles');
     Route::delete('/users/{user}/roles/{role}', [UserController::class, 'removeRole'])->name('users.roles.remove');
-    Route::post('/users/{user}/permissions', [UserController::class, 'givePermission'])->name('users.permissions');
-    Route::delete('/users/{user}/permissions/{permission}', [UserController::class, 'revokePermission'])->name('users.permissions.revoke');
+    Route::post('/users/{user}/permission/permissions', [UserController::class, 'givePermission'])->name('users.permissions');
+    Route::delete('/users/{user}/permission/permissions/{permission}', [UserController::class, 'revokePermission'])->name('users.permissions.revoke');
 
 });
